Created working audio transcription.
Eventually will want to reuse some of it inside
of the other function to transcribe+summarize from chatgpt. otherwise, I will also want to do stuff like probably
set up a system to expect specific outputs from
chatgpt. ie wanting it to return in as a json file.
This could use some more overhead. but otherwise, it will be useful for that stuff...