Database DOES actually work, explaination:
Right on school networks the port auto blocks the supabase port. this makes it so that it is actually inaccessible to the computer when running locally. So beyond that. we need to make sure that things will be working on the port etc so that this is all actually real and working.